+++ dotnet/src/SmartStandards.Logging/Logging/SmartStandards/Transport/TraceBusFeed.Buffer.cs
@@ -5,143 +5,63 @@ namespace Logging.SmartStandards.Transport {
 
   public partial class TraceBusFeed {
 
+    /// <remarks>
+    ///   Taken from https://stackoverflow.com/a/5924776
+    /// </remarks>
     private class MyCircularBuffer<T> : IEnumerable<T> {
 
-      readonly int size;
+      readonly int _Size;
 
-      readonly object locker;
+      readonly object _Locker;
 
-      int count;
-      int head;
-      int rear;
-      T[] values;
+      int _Count;
+      int _Head;
+      int _Rear;
+      T[] _Values;
 
       public MyCircularBuffer(int max) {
-        this.size = max;
-        locker = new object();
-        count = 0;
-        head = 0;
-        rear = 0;
-        values = new T[size];
+        _Size = max;
+        _Locker = new object();
+        _Count = 0;
+        _Head = 0;
+        _Rear = 0;
+        _Values = new T[_Size];
       }
 
       static int Incr(int index, int size) {
         return (index + 1) % size;
       }
 
-      private void UnsafeEnsureQueueNotEmpty() {
-        if (count == 0) throw new Exception("Empty queue");
-      }
-
-      public int Size { get { return size; } }
-      public object SyncRoot { get { return locker; } }
-
-      #region Count
-
-      public int Count { get { return UnsafeCount; } }
-      public int SafeCount { get { lock (locker) { return UnsafeCount; } } }
-      public int UnsafeCount { get { return count; } }
-
-      #endregion
-
-      #region Enqueue
-
-      public void Enqueue(T obj) {
-        UnsafeEnqueue(obj);
-      }
+      public object SyncRoot { get { return _Locker; } }
 
       public void SafeEnqueue(T obj) {
-        lock (locker) { UnsafeEnqueue(obj); }
+        lock (_Locker) { UnsafeEnqueue(obj); }
       }
 
       public void UnsafeEnqueue(T obj) {
 
-        values[rear] = obj;
+        _Values[_Rear] = obj;
 
-        if (Count == Size) head = Incr(head, Size);
+        if (_Count == _Size) _Head = Incr(_Head, _Size);
 
-        rear = Incr(rear, Size);
-        count = Math.Min(count + 1, Size);
+        _Rear = Incr(_Rear, _Size);
+        _Count = Math.Min(_Count + 1, _Size);
       }
-
-      #endregion
-
-      #region Dequeue
-
-      public T Dequeue() {
-        return UnsafeDequeue();
-      }
-
-      public T SafeDequeue() {
-        lock (locker) { return UnsafeDequeue(); }
-      }
-
-      public T UnsafeDequeue() {
-        UnsafeEnsureQueueNotEmpty();
-
-        T res = values[head];
-        values[head] = default(T);
-        head = Incr(head, Size);
-        count--;
-
-        return res;
-      }
-
-      #endregion
-
-      #region Peek
-
-      public T Peek() {
-        return UnsafePeek();
-      }
-
-      public T SafePeek() {
-        lock (locker) { return UnsafePeek(); }
-      }
-
-      public T UnsafePeek() {
-        UnsafeEnsureQueueNotEmpty();
-
-        return values[head];
-      }
-
-      #endregion
-
-
-      #region GetEnumerator
 
       public IEnumerator<T> GetEnumerator() {
-        return UnsafeGetEnumerator();
-      }
+        int index = _Head;
 
-      public IEnumerator<T> SafeGetEnumerator() {
-        lock (locker) {
-
-          List<T> res = new List<T>(count);
-
-          var enumerator = UnsafeGetEnumerator();
-
-          while (enumerator.MoveNext()) res.Add(enumerator.Current);
-
-          return res.GetEnumerator();
+        for (int i = 0; i < _Count; i++) {
+          yield return _Values[index];
+          index = Incr(index, _Size);
         }
-      }
 
-      public IEnumerator<T> UnsafeGetEnumerator() {
-
-        int index = head;
-
-        for (int i = 0; i < count; i++) {
-          yield return values[index];
-          index = Incr(index, size);
-        }
       }
 
       System.Collections.IEnumerator System.Collections.IEnumerable.GetEnumerator() {
         return this.GetEnumerator();
       }
 
-      #endregion
     }
 
   }