+++ dotnet/test/SmartStandards.Logging.ThirdPartyTests/AssemblyInitializer.cs
@@ -1,0 +1,15 @@
+﻿using Logging.SmartStandards.Transport;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+[TestClass]
+public class AssemblyInitializer {
+
+  [AssemblyInitialize]
+  public static void InitializeAssembly(TestContext testContext) {
+
+    TraceBusFeed.Instance.RequiredListeners.Add("SmartStandards395316649");
+
+  }
+
+}
+