+++ vers/changelog.md
@@ -5,9 +5,7 @@ This files contains a version history including all changes relevant for semanti
 
 ## Upcoming Changes
 
-*(none)*
-
-
+- **new Feature**: TraceBusFeed: Can now be used as singleton.
 
 ## v 2.1.0
 released **2025-04-09**, including: